Rearrange context extensions
